requirements.txt
json
numpy
PyQt5
Pillow
opencv-python
matplotlib
